PPTP-1550 Clean up; move partner contact name, email and phone number views to the partner package.
This is where other partners code has converged to.
